requirements.txt
asgiref==3.7.2
Django==4.2.9
django-appconf==1.0.6
django-ckeditor==5.9.0
django-imagekit==4.1.0
django-js-asset==2.2.0
django-tinymce==3.7.1
pilkit==3.0
Pillow==9.2.0
python-decouple==3.8
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
